refactor(tests): organize interpreter stdlib collections tests
