+++ crates/rl-tests/tests/interpreter/stdlib/collections/mod.rs
@@ -1,0 +1,2 @@
+mod sets;
+mod maps;
